refactored task item components
